+++ praw/models/listing/mixins/subreddit.py
@@ -37,7 +37,7 @@ class CommentHelper(PRAWBase):
 
         .. code-block:: python
 
-            for comment in reddit.subreddit("redditdev").comments(limit=25):
+            for comment in reddit.subreddit("test").comments(limit=25):
                 print(comment.author)
 
         """